+++ sources/packages/load-test/src/utils/auth/models.ts
@@ -13,3 +13,13 @@ export interface TokenCacheResponse {
   accessToken: string;
   expiresIn: number;
 }
+
+/**
+ * Authorized parties recognized by the API.
+ */
+export enum AuthorizedParties {
+  Institution = "institution",
+  Student = "student",
+  AEST = "aest",
+  SupportingUsers = "supporting-users",
+}
